requirements.txt
discord.py
PyNaCl
jishaku
aiosqlite
python-dotenv
youtube-dl
